TW-2928 Reorder contacts list
