+++ lib/pages/contacts_tab/contacts_tab_body_view.dart
@@ -1,24 +1,16 @@
-import 'package:fluffychat/domain/app_state/contact/get_contacts_state.dart';
 import 'package:fluffychat/pages/contacts_tab/contacts_tab.dart';
 import 'package:fluffychat/pages/contacts_tab/contacts_tab_view_style.dart';
-import 'package:fluffychat/pages/contacts_tab/empty_contacts_body.dart';
+import 'package:fluffychat/pages/contacts_tab/widgets/sliver_contacts_with_matrix_id.dart';
+import 'package:fluffychat/pages/contacts_tab/widgets/sliver_contacts_without_matrix_id.dart';
+import 'package:fluffychat/pages/contacts_tab/widgets/sliver_empty_contacts.dart';
 import 'package:fluffychat/pages/contacts_tab/widgets/sliver_invite_friend_button.dart';
-import 'package:fluffychat/pages/new_private_chat/widget/expansion_contact_list_tile.dart';
-import 'package:fluffychat/pages/new_private_chat/widget/expansion_phonebook_contact_list_tile.dart';
-import 'package:fluffychat/pages/new_private_chat/widget/loading_contact_widget.dart';
-import 'package:fluffychat/pages/new_private_chat/widget/no_contacts_found.dart';
-import 'package:fluffychat/pages/search/recent_item_widget.dart';
-import 'package:fluffychat/presentation/model/contact/get_presentation_contacts_empty.dart';
-import 'package:fluffychat/presentation/model/contact/get_presentation_contacts_failure.dart';
-import 'package:fluffychat/presentation/model/contact/presentation_contact.dart';
-import 'package:fluffychat/presentation/model/contact/presentation_contact_success.dart';
-import 'package:fluffychat/presentation/model/search/presentation_search.dart';
+import 'package:fluffychat/pages/contacts_tab/widgets/sliver_loading_contacts.dart';
+import 'package:fluffychat/pages/contacts_tab/widgets/sliver_phonebook_contacts_with_matrix_id.dart';
+import 'package:fluffychat/pages/contacts_tab/widgets/sliver_phonebook_contacts_without_matrix_id.dart';
+import 'package:fluffychat/pages/contacts_tab/widgets/sliver_recent_contacts.dart';
+import 'package:fluffychat/pages/contacts_tab/widgets/sliver_warning_banner.dart';
 import 'package:fluffychat/utils/platform_infos.dart';
-import 'package:fluffychat/widgets/contacts_warning_banner/contacts_warning_banner_view.dart';
-import 'package:fluffychat/widgets/sliver_expandable_list.dart';
 import 'package:flutter/material.dart';
-import 'package:fluffychat/generated/l10n/app_localizations.dart';
-import 'package:matrix/matrix.dart';
 
 class ContactsTabBodyView extends StatelessWidget {
   final ContactsTabController controller;
@@ -31,461 +23,20 @@ class ContactsTabBodyView extends StatelessWidget {
       slivers: [
         if (controller.client.userID != null)
           SliverInviteFriendButton(userId: controller.client.userID!),
-        _SliverWarningBanner(controller: controller),
-        _SliverRecentContacts(controller: controller),
-        _SliverContactsList(controller: controller),
+        SliverWarningBanner(controller: controller),
+        SliverRecentContacts(controller: controller),
+        SliverLoadingContacts(controller: controller),
+        SliverContactsWithMatrixId(controller: controller),
         if (PlatformInfos.isMobile)
-          _SliverPhonebookList(controller: controller),
-        const _SliverPadding(),
+          SliverPhonebookContactsWithMatrixId(controller: controller),
+        SliverContactsWithoutMatrixId(controller: controller),
+        if (PlatformInfos.isMobile)
+          SliverPhonebookContactsWithoutMatrixId(controller: controller),
+        SliverEmptyContacts(controller: controller),
+        const SliverToBoxAdapter(
+          child: SizedBox(height: ContactsTabViewStyle.padding),
+        ),
       ],
-    );
-  }
-}
-
-class _SliverPadding extends StatelessWidget {
-  const _SliverPadding();
-
-  @override
-  Widget build(BuildContext context) {
-    return const SliverToBoxAdapter(
-      child: SizedBox(height: ContactsTabViewStyle.padding),
-    );
-  }
-}
-
-class _SliverPhonebookList extends StatelessWidget {
-  const _SliverPhonebookList({required this.controller});
-
-  final ContactsTabController controller;
-
-  @override
-  Widget build(BuildContext context) {
-    return ValueListenableBuilder(
-      valueListenable: controller.presentationPhonebookContactNotifier,
-      builder: (context, phonebookContactState, child) {
-        return phonebookContactState.fold(
-          (failure) {
-            if (!PlatformInfos.isMobile) {
-              return child!;
-            }
-            final presentationRecentContact =
-                controller.presentationRecentContactNotifier.value;
-            if (failure is GetPresentationContactsFailure) {
-              if (presentationRecentContact.isEmpty) {
-                return controller.presentationContactNotifier.value.fold((
-                  failure,
-                ) {
-                  if (failure is GetPresentationContactsFailure ||
-                      failure is GetPresentationContactsEmpty) {
-                    return SliverToBoxAdapter(
-                      child: Padding(
-                        padding: const EdgeInsets.only(
-                          left: ContactsTabViewStyle.padding,
-                          top: ContactsTabViewStyle.padding,
-                        ),
-                        child: NoContactsFound(
-                          keyword: controller.textEditingController.text.isEmpty
-                              ? null
-                              : controller.textEditingController.text,
-                        ),
-                      ),
-                    );
-                  }
-                  return child!;
-                }, (_) => child!);
-              }
-            }
-            if (failure is GetPresentationContactsEmpty) {
-              if (presentationRecentContact.isEmpty) {
-                return controller.presentationContactNotifier.value.fold((
-                  failure,
-                ) {
-                  if (failure is GetPresentationContactsFailure ||
-                      failure is GetPresentationContactsEmpty) {
-                    if (controller.textEditingController.text.isEmpty) {
-                      return const SliverToBoxAdapter(
-                        child: EmptyContactBody(),
-                      );
-                    } else {
-                      return SliverToBoxAdapter(
-                        child: Padding(
-                          padding: const EdgeInsets.only(
-                            left: ContactsTabViewStyle.padding,
-                            top: ContactsTabViewStyle.padding,
-                          ),
-                          child: NoContactsFound(
-                            keyword: controller.textEditingController.text,
-                          ),
-                        ),
-                      );
-                    }
-                  }
-                  return child!;
-                }, (_) => child!);
-              }
-            }
-            return child!;
-          },
-          (success) {
-            if (success is PresentationContactsSuccess) {
-              final contacts = success.contacts;
-
-              return SliverExpandableList(
-                title: L10n.of(context)!.contactsCount(contacts.length),
-                itemCount: contacts.length,
-                itemBuilder: (context, index) => _PhonebookContact(
-                  contact: contacts[index],
-                  controller: controller,
-                ),
-              );
-            }
-            return child!;
-          },
-        );
-      },
-      child: const SliverToBoxAdapter(child: SizedBox()),
-    );
-  }
-}
-
-class _SliverContactsList extends StatelessWidget {
-  const _SliverContactsList({required this.controller});
-
-  final ContactsTabController controller;
-
-  @override
-  Widget build(BuildContext context) {
-    return ValueListenableBuilder(
-      valueListenable: controller.presentationContactNotifier,
-      builder: (context, state, child) {
-        return state.fold(
-          (failure) {
-            if (PlatformInfos.isMobile) {
-              return child!;
-            }
-            final presentationRecentContact =
-                controller.presentationRecentContactNotifier.value;
-
-            if (presentationRecentContact.isNotEmpty) {
-              return child!;
-            }
-            if (PlatformInfos.isWeb) {
-              if (failure is GetPresentationContactsFailure ||
-                  failure is GetPresentationContactsEmpty) {
-                final keyword = controller.textEditingController.text;
-                if (keyword.isEmpty) {
-                  return const SliverToBoxAdapter(child: EmptyContactBody());
-                } else {
-                  return SliverToBoxAdapter(
-                    child: Padding(
-                      padding: const EdgeInsets.only(
-                        left: ContactsTabViewStyle.padding,
-                        top: ContactsTabViewStyle.padding,
-                      ),
-                      child: NoContactsFound(
-                        keyword: controller.textEditingController.text,
-                      ),
-                    ),
-                  );
-                }
-              }
-              return child!;
-            } else {
-              return controller.presentationPhonebookContactNotifier.value.fold(
-                (_) {
-                  if (controller.presentationPhonebookContactNotifier.value
-                      .isRight()) {
-                    return child!;
-                  }
-                  if (failure is GetPresentationContactsFailure ||
-                      failure is GetPresentationContactsEmpty) {
-                    final keyword = controller.textEditingController.text;
-                    if (keyword.isEmpty) {
-                      return const SliverToBoxAdapter(
-                        child: EmptyContactBody(),
-                      );
-                    } else {
-                      return SliverToBoxAdapter(
-                        child: Padding(
-                          padding: const EdgeInsets.only(
-                            left: ContactsTabViewStyle.padding,
-                            top: ContactsTabViewStyle.padding,
-                          ),
-                          child: NoContactsFound(
-                            keyword: controller.textEditingController.text,
-                          ),
-                        ),
-                      );
-                    }
-                  }
-                  return child!;
-                },
-                (success) => child!,
-              );
-            }
-          },
-          (success) {
-            if (success is ContactsLoading) {
-              return const SliverToBoxAdapter(child: LoadingContactWidget());
-            }
-
-            if (success is PresentationExternalContactSuccess) {
-              if (controller
-                  .presentationRecentContactNotifier
-                  .value
-                  .isNotEmpty) {
-                return child!;
-              }
-
-              if (!PlatformInfos.isWeb) {
-                if (controller.phoneBookFilterSuccess) {
-                  return child!;
-                }
-              }
-
-              final externalContact = success.contact;
-              return _SilverExternalContact(
-                controller: controller,
-                externalContact: externalContact,
-              );
-            }
-
-            if (success is PresentationContactsSuccess) {
-              final contacts = success.contacts;
-              return SliverExpandableList(
-                title: L10n.of(context)!.linagoraContactsCount(contacts.length),
-                itemCount: contacts.length,
-                itemBuilder: (context, index) =>
-                    _Contact(contact: contacts[index], controller: controller),
-              );
-            }
-
-            return child!;
-          },
-        );
-      },
-      child: const SliverToBoxAdapter(child: SizedBox()),
-    );
-  }
-}
-
-class _SilverExternalContact extends StatefulWidget {
-  final ContactsTabController controller;
-  final PresentationContact externalContact;
-
-  const _SilverExternalContact({
-    required this.controller,
-    required this.externalContact,
-  });
-
-  @override
-  State<_SilverExternalContact> createState() => _SilverExternalContactState();
-}
-
-class _SilverExternalContactState extends State<_SilverExternalContact> {
-  Future<CachedProfileInformation>? _profileFuture;
-  String? _currentMatrixId;
-
-  @override
-  void didChangeDependencies() {
-    super.didChangeDependencies();
-    _fetchProfileIfNeeded();
-  }
-
-  @override
-  void didUpdateWidget(covariant _SilverExternalContact oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    if (oldWidget.externalContact.matrixId != widget.externalContact.matrixId) {
-      _fetchProfileIfNeeded();
-    }
-  }
-
-  void _fetchProfileIfNeeded() {
-    final matrixId = widget.externalContact.matrixId;
-    if (matrixId == null) {
-      _currentMatrixId = null;
-      _profileFuture = null;
-      return;
-    }
-    if (_currentMatrixId != matrixId) {
-      _currentMatrixId = matrixId;
-      _profileFuture = widget.controller.client.getUserProfile(
-        matrixId,
-        maxCacheAge: Duration.zero,
-      );
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return FutureBuilder<CachedProfileInformation>(
-      future: _profileFuture,
-      builder: (context, snapshot) {
-        if (snapshot.connectionState == ConnectionState.waiting) {
-          return const SliverToBoxAdapter(child: LoadingContactWidget());
-        }
-
-        if (snapshot.hasError || !snapshot.hasData) {
-          return SliverToBoxAdapter(
-            child: Padding(
-              padding: const EdgeInsets.only(
-                left: ContactsTabViewStyle.padding,
-                top: ContactsTabViewStyle.padding,
-              ),
-              child: NoContactsFound(
-                keyword: widget.controller.textEditingController.text,
-              ),
-            ),
-          );
-        }
-
-        final profile = snapshot.data!;
-        final validatedContact = PresentationContact(
-          matrixId: widget.externalContact.matrixId,
-          displayName:
-              profile.displayname ?? widget.externalContact.displayName,
-          type: widget.externalContact.type,
-        );
-
-        return SliverToBoxAdapter(
-          child: Padding(
-            padding: const EdgeInsets.symmetric(
-              horizontal: ContactsTabViewStyle.padding,
-            ),
-            child: ExpansionContactListTile(
-              contact: validatedContact,
-              highlightKeyword: widget.controller.textEditingController.text,
-              enableInvitation: widget.controller.supportInvitation(),
-              onContactTap: () => widget.controller.onContactTap(
-                context: context,
-                path: 'rooms',
-                contact: validatedContact,
-              ),
-            ),
-          ),
-        );
-      },
-    );
-  }
-}
-
-class _SliverRecentContacts extends StatelessWidget {
-  final ContactsTabController controller;
-
-  const _SliverRecentContacts({required this.controller});
-
-  @override
-  Widget build(BuildContext context) {
-    return ValueListenableBuilder(
-      valueListenable: controller.presentationContactNotifier,
-      builder: (context, state, child) {
-        return state.fold((failure) => child!, (success) {
-          if (success is ContactsLoading) {
-            return const SliverToBoxAdapter(child: SizedBox());
-          }
-          return child!;
-        });
-      },
-      child: ValueListenableBuilder(
-        valueListenable: controller.presentationRecentContactNotifier,
-        builder: (context, recentContacts, child) {
-          if (recentContacts.isEmpty) {
-            return child!;
-          }
-          return SliverExpandableList(
-            title: L10n.of(context)!.recent,
-            itemCount: recentContacts.length,
-            itemBuilder: (context, index) => Padding(
-              padding: const EdgeInsets.symmetric(
-                horizontal: ContactsTabViewStyle.padding,
-              ),
-              child: RecentItemWidget(
-                presentationSearch: recentContacts[index],
-                highlightKeyword: controller.textEditingController.text,
-                client: controller.client,
-                key: Key('contact_recent_${recentContacts[index].id}'),
-                onTap: () => controller.onContactTap(
-                  contact: recentContacts[index].toPresentationContact(),
-                  context: context,
-                  path: 'rooms',
-                ),
-                avatarSize: ContactsTabViewStyle.avatarSize,
-              ),
-            ),
-          );
-        },
-        child: const SliverToBoxAdapter(child: SizedBox()),
-      ),
-    );
-  }
-}
-
-class _SliverWarningBanner extends StatelessWidget {
-  const _SliverWarningBanner({required this.controller});
-
-  final ContactsTabController controller;
-
-  @override
-  Widget build(BuildContext context) {
-    return SliverToBoxAdapter(
-      child: ContactsWarningBannerView(
-        warningBannerNotifier: controller.warningBannerNotifier,
-        closeContactsWarningBanner: controller.closeContactsWarningBanner,
-        goToSettingsForPermissionActions: () =>
-            controller.displayContactPermissionDialog(context),
-      ),
-    );
-  }
-}
-
-class _Contact extends StatelessWidget {
-  const _Contact({required this.contact, required this.controller});
-
-  final PresentationContact contact;
-  final ContactsTabController controller;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(
-        horizontal: ContactsTabViewStyle.padding,
-      ),
-      child: ExpansionContactListTile(
-        contact: contact,
-        highlightKeyword: controller.textEditingController.text,
-        enableInvitation: controller.supportInvitation(),
-        onContactTap: () => controller.onContactTap(
-          context: context,
-          path: 'rooms',
-          contact: contact,
-        ),
-      ),
-    );
-  }
-}
-
-class _PhonebookContact extends StatelessWidget {
-  const _PhonebookContact({required this.contact, required this.controller});
-
-  final PresentationContact contact;
-  final ContactsTabController controller;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(
-        horizontal: ContactsTabViewStyle.padding,
-      ),
-      child: ExpansionPhonebookContactListTile(
-        contact: contact,
-        highlightKeyword: controller.textEditingController.text,
-        enableInvitation: controller.supportInvitation(),
-        onContactTap: () => controller.onContactTap(
-          context: context,
-          path: 'rooms',
-          contact: contact,
-        ),
-      ),
     );
   }
 }
